apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
python-multipart==0.0.20
python-dateutil==2.9.0.post0
pytest==8.4.1
